move env to its own file, remove serialization
