+++ src/configuration/mod.rs
@@ -1,9 +1,13 @@
 //! Configuration module
 
+mod env;
+
 use crate::error::Error as PrimaDatadogError;
 use crate::TagTrackerConfiguration;
 use std::fmt::Display;
 use std::str::FromStr;
+
+use self::env::Environment;
 
 /// By binding to 0.0.0.0:0 we're just letting the OS assign us a port, and letting anyone send us UDP packets on that port
 ///
@@ -108,82 +112,6 @@ impl From<Configuration> for dogstatsd::Options {
             value.socket_path(),
             value.batching_options(),
         )
-    }
-}
-
-/// Represent an environment in which the datadog client runs.
-/// This is useful for enforcing rules based on environment for every application that uses the library.
-#[derive(PartialEq, Eq, Debug, Clone)]
-pub enum Environment {
-    Dev,
-    Staging,
-    Production,
-}
-
-impl Environment {
-    /// Returns the string representation of the environment.
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Environment::Dev => "dev",
-            Environment::Staging => "staging",
-            Environment::Production => "production",
-        }
-    }
-}
-
-impl FromStr for Environment {
-    type Err = PrimaDatadogError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "dev" => Ok(Self::Dev),
-            "staging" => Ok(Self::Staging),
-            "production" => Ok(Self::Production),
-            _ => Err(PrimaDatadogError::WrongEnvironmentDefinition),
-        }
-    }
-}
-
-impl Display for Environment {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-#[cfg_attr(docsrs, doc(cfg(feature = "serde")))]
-#[cfg(feature = "serde")]
-impl serde::Serialize for Environment {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        serializer.serialize_str(self.as_str())
-    }
-}
-
-#[cfg_attr(docsrs, doc(cfg(feature = "serde")))]
-#[cfg(feature = "serde")]
-impl<'de> serde::Deserialize<'de> for Environment {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        struct EnvironmentVisitor;
-        impl<'de> serde::de::Visitor<'de> for EnvironmentVisitor {
-            type Value = Environment;
-
-            fn expecting(&self, formatter: &mut std::fmt::Formatter) -> std::fmt::Result {
-                formatter.write_str("a environment tag")
-            }
-
-            fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                Environment::from_str(v).map_err(|_| E::custom("unknown environment tag"))
-            }
-        }
-        deserializer.deserialize_str(EnvironmentVisitor)
     }
 }
 
@@ -295,15 +223,6 @@ mod tests {
     }
 
     #[test]
-    pub fn test_environment() {
-        let config = Configuration::new("to_addr", "namespace")
-            .with_from_addr("from_addr")
-            .with_environment(Environment::Dev);
-
-        assert_eq!(config.default_tags(), vec!["env:dev"]);
-    }
-
-    #[test]
     pub fn test_country() {
         let config = Configuration::new("to_addr", "namespace")
             .with_from_addr("from_addr")
@@ -327,16 +246,6 @@ mod tests {
             let serialized = serde_json::to_string(&country).unwrap();
             let deserialized: Country = serde_json::from_str(&serialized).unwrap();
             assert_eq!(country, deserialized);
-        }
-    }
-
-    #[cfg(feature = "serde")]
-    #[test]
-    fn test_environment_serde() {
-        for environment in [Environment::Production, Environment::Staging, Environment::Dev] {
-            let serialized = serde_json::to_string(&environment).unwrap();
-            let deserialized: Environment = serde_json::from_str(&serialized).unwrap();
-            assert_eq!(environment, deserialized);
         }
     }
 }
